# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Adapta API
  description: |
    **Your model. Your data. Your servers.**

    Adapta is a self-hosted model-customization platform. Technical teams deploy
    it on their own infrastructure to specialize and serve private language models two ways,
    each behind an OpenAI-compatible API. Your data never leaves your servers.

    **Give it knowledge (RAG)** — upload documents, embed them, and get answers grounded in
    your content, with citations. Weights never change. CPU-only.

    **Change how it behaves (fine-tuning / LoRA)** — train a QLoRA adapter on instruction
    data; the adapter only serves once it passes the eval gate (score ≥ 0.6 or a clear
    improvement over the base model). Requires a GPU.

    **Composed endpoints** — a single project endpoint can serve both: retrieved context
    with citations (RAG) and trained tone/format (the LoRA adapter) in one call.

    ---

    **Auth:** management endpoints require a JWT Bearer token from `POST /v1/auth/login`.
    The `POST /v1/chat/completions` endpoint uses a scoped project API key (`adp_*` prefix).

  version: 0.1.0
  license:
    name: MIT
    url: https://github.com/yielab/adapta/blob/main/LICENSE
  contact:
    name: Adapta
    url: https://github.com/yielab/adapta
  x-logo:
    url: /console/logo.svg
    altText: Adapta
servers:
  - url: /v1
    description: Versioned API
...
paths:
...
  /projects/{project_id}/files:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a document for indexing (both project types)
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: {type: string, format: binary}
              required: [file]
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '202': {description: Accepted — indexing in background}

    get:
      operationId: listFiles
      tags: [files]
      summary: List uploaded files for a project
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Files, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/FileResponse'}}}}}
...
  /projects/{project_id}/datasets:
    post:
      operationId: uploadDataset
      tags: [datasets]
      summary: Upload a training dataset (finetune projects) — .jsonl, or a .zip image bundle
      description: |
        Plain `.jsonl` for text datasets. For image datasets (§V), a `.zip`
        bundle holding exactly one root `.jsonl` manifest plus the images its
        rows reference (`images` paths relative to the bundle; png/jpg/jpeg/webp).
        Validation runs in the background; poll the dataset until `valid`/`invalid`.
        Bundles are extracted with traversal protection and size/count caps.
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: {type: string, format: binary}
              required: [file]
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '202': {description: Accepted — validation in background}

    get:
      operationId: listDatasets
      tags: [datasets]
      summary: List datasets for a project
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Datasets, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/DatasetResponse'}}}}}
  /projects/{project_id}/datasets/{dataset_id}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get dataset status
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ProjectId'}
        - {name: dataset_id, in: path, required: true, schema: {type: string}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Dataset, content: {application/json: {schema: {$ref: '#/components/schemas/DatasetResponse'}}}}

  # ──────────────────────────────────────────────────────────────────────────
  # Training jobs
  # ──────────────────────────────────────────────────────────────────────────
...
  /projects/{project_id}/endpoint:
    post:
      operationId: createEndpoint
      tags: [endpoints]
      summary: Activate the project endpoint (RAG must have indexed docs; finetune must have passed eval gate)
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '201': {description: Endpoint created, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointResponse'}}}}
        '422': {$ref: '#/components/responses/Error'}

    get:
      operationId: getEndpoint
      tags: [endpoints]
      summary: Get project endpoint
      security: [{BearerAuth: []}]
      parameters: [{$ref: '#/components/parameters/ProjectId'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '200': {description: Endpoint, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointResponse'}}}}

  # ──────────────────────────────────────────────────────────────────────────
  # Usage metering
  # ──────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: Scoped project API key (adp_...)
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: {type: string}
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    ProjectType:
      type: string
      enum: [rag, finetune]
    Modality:
      type: string
      enum: [text, vision]
    Gate:
      type: string
      enum: [absolute, improvement]
...
    EndpointStatus:
      type: string
      enum: [pending, active, disabled]
...
    FileResponse:
      type: object
      required: [id, filename, status, num_chunks, size_bytes, uploaded_at]
      properties:
        id: {type: string}
        filename: {type: string}
        status: {type: string}
        num_chunks: {type: integer, nullable: true}
        size_bytes: {type: integer}
        uploaded_at: {type: string}

    # Datasets
    DatasetResponse:
      type: object
      required: [id, name, status, num_samples, modality, num_images, validation_error, created_at]
      properties:
        id: {type: string}
        name: {type: string}
        status: {type: string}
        num_samples: {type: integer, nullable: true}
        # §V image datasets: "vision" once an image bundle validates; "text" otherwise.
        modality: {$ref: '#/components/schemas/Modality'}
        num_images: {type: integer, nullable: true}
        validation_error: {type: string, nullable: true}
        created_at: {type: string}

    # Jobs
...
    AdapterProvenance:
      type: object
      description: Adapter eval evidence (null for RAG-only endpoints)
      required: [job_id, eval_score, gate]
      properties:
        job_id: {type: string}
        eval_score: {type: number, format: float}
        base_score: {type: number, format: float, nullable: true}
        score_delta: {type: number, format: float, nullable: true}
        gate: {$ref: '#/components/schemas/Gate'}
    RetrievalSummary:
      type: object
      description: RAG retrieval layer (null when no documents are indexed)
      required: [indexed_chunks]
      properties:
        indexed_chunks: {type: integer}
    EndpointResponse:
      type: object
      required: [id, slug, status, base_model, modality, project_type, created_at, adapter, retrieval, adapter_path]
      properties:
        id: {type: string}
        slug: {type: string}
        status: {$ref: '#/components/schemas/EndpointStatus'}
        base_model: {type: string}
        modality: {$ref: '#/components/schemas/Modality'}
        project_type: {$ref: '#/components/schemas/ProjectType'}
        created_at: {type: string}
        adapter: {$ref: '#/components/schemas/AdapterProvenance', nullable: true}
        retrieval: {$ref: '#/components/schemas/RetrievalSummary', nullable: true}
        adapter_path: {type: string, nullable: true, deprecated: true}

    # Usage
...
    Error:
      type: object
      properties:
        code: {type: string}
        message: {type: string}
        correlation_id: {type: string, nullable: true}
    ErrorResponse:
      type: object
      properties:
        error: {$ref: '#/components/schemas/Error'}